feat: core payment package
