+++ core/data.go
@@ -4,18 +4,10 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"math/big"
-	"strconv"
-	"time"
 
-	commonpbv2 "github.com/Layr-Labs/eigenda/api/grpc/common/v2"
 	"github.com/Layr-Labs/eigenda/common"
 	"github.com/Layr-Labs/eigenda/encoding"
-	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/consensys/gnark-crypto/ecc/bn254"
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	gethcommon "github.com/ethereum/go-ethereum/common"
-	"golang.org/x/crypto/sha3"
 )
 
 type AccountID = string
@@ -494,182 +486,6 @@ func (cb Bundles) FromEncodedBundles(eb EncodedBundles) (Bundles, error) {
 	return c, nil
 }
 
-// PaymentMetadata represents the header information for a blob
-type PaymentMetadata struct {
-	// AccountID is the ETH account address for the payer
-	AccountID gethcommon.Address `json:"account_id"`
-
-	// Timestamp represents the nanosecond of the dispersal request creation
-	Timestamp int64 `json:"timestamp"`
-	// CumulativePayment represents the total amount of payment (in wei) made by the user up to this point
-	CumulativePayment *big.Int `json:"cumulative_payment"`
-}
-
-// Hash returns the Keccak256 hash of the PaymentMetadata
-func (pm *PaymentMetadata) Hash() ([32]byte, error) {
-	if pm == nil {
-		return [32]byte{}, errors.New("payment metadata is nil")
-	}
-	blobHeaderType, err := abi.NewType("tuple", "", []abi.ArgumentMarshaling{
-		{
-			Name: "accountID",
-			Type: "string",
-		},
-		{
-			Name: "timestamp",
-			Type: "int64",
-		},
-		{
-			Name: "cumulativePayment",
-			Type: "uint256",
-		},
-	})
-	if err != nil {
-		return [32]byte{}, err
-	}
-
-	arguments := abi.Arguments{
-		{
-			Type: blobHeaderType,
-		},
-	}
-
-	s := struct {
-		AccountID         string
-		Timestamp         int64
-		CumulativePayment *big.Int
-	}{
-		AccountID:         pm.AccountID.Hex(),
-		Timestamp:         pm.Timestamp,
-		CumulativePayment: pm.CumulativePayment,
-	}
-
-	bytes, err := arguments.Pack(s)
-	if err != nil {
-		return [32]byte{}, err
-	}
-
-	var hash [32]byte
-	hasher := sha3.NewLegacyKeccak256()
-	hasher.Write(bytes)
-	copy(hash[:], hasher.Sum(nil)[:32])
-
-	return hash, nil
-}
-
-func (pm *PaymentMetadata) MarshalDynamoDBAttributeValue() (types.AttributeValue, error) {
-	if pm == nil {
-		return nil, errors.New("payment metadata is nil")
-	}
-
-	return &types.AttributeValueMemberM{
-		Value: map[string]types.AttributeValue{
-			"AccountID": &types.AttributeValueMemberS{Value: pm.AccountID.Hex()},
-			"Timestamp": &types.AttributeValueMemberN{Value: fmt.Sprintf("%d", pm.Timestamp)},
-			"CumulativePayment": &types.AttributeValueMemberN{
-				Value: pm.CumulativePayment.String(),
-			},
-		},
-	}, nil
-}
-
-func (pm *PaymentMetadata) UnmarshalDynamoDBAttributeValue(av types.AttributeValue) error {
-	m, ok := av.(*types.AttributeValueMemberM)
-	if !ok {
-		return fmt.Errorf("expected *types.AttributeValueMemberM, got %T", av)
-	}
-	accountID, ok := m.Value["AccountID"].(*types.AttributeValueMemberS)
-	if !ok {
-		return fmt.Errorf("expected *types.AttributeValueMemberS for AccountID, got %T", m.Value["AccountID"])
-	}
-	pm.AccountID = gethcommon.HexToAddress(accountID.Value)
-	rp, ok := m.Value["Timestamp"].(*types.AttributeValueMemberN)
-	if !ok {
-		return fmt.Errorf("expected *types.AttributeValueMemberN for Timestamp, got %T", m.Value["Timestamp"])
-	}
-	timestamp, err := strconv.ParseInt(rp.Value, 10, 64)
-	if err != nil {
-		return fmt.Errorf("failed to parse Timestamp: %w", err)
-	}
-	pm.Timestamp = timestamp
-	cp, ok := m.Value["CumulativePayment"].(*types.AttributeValueMemberN)
-	if !ok {
-		return fmt.Errorf("expected *types.AttributeValueMemberN for CumulativePayment, got %T", m.Value["CumulativePayment"])
-	}
-	pm.CumulativePayment, _ = new(big.Int).SetString(cp.Value, 10)
-	return nil
-}
-
-func (pm *PaymentMetadata) ToProtobuf() *commonpbv2.PaymentHeader {
-	if pm == nil {
-		return nil
-	}
-	return &commonpbv2.PaymentHeader{
-		AccountId:         pm.AccountID.Hex(),
-		Timestamp:         pm.Timestamp,
-		CumulativePayment: pm.CumulativePayment.Bytes(),
-	}
-}
-
-// ConvertToProtoPaymentHeader converts a PaymentMetadata to a protobuf payment header
-func ConvertToPaymentMetadata(ph *commonpbv2.PaymentHeader) (*PaymentMetadata, error) {
-	if ph == nil {
-		return nil, nil
-	}
-
-	if !gethcommon.IsHexAddress(ph.AccountId) {
-		return nil, fmt.Errorf("invalid account ID: %s", ph.AccountId)
-	}
-
-	return &PaymentMetadata{
-		AccountID:         gethcommon.HexToAddress(ph.AccountId),
-		Timestamp:         ph.Timestamp,
-		CumulativePayment: new(big.Int).SetBytes(ph.CumulativePayment),
-	}, nil
-}
-
-// ReservedPayment contains information the onchain state about a reserved payment
-type ReservedPayment struct {
-	// reserve number of symbols per second
-	SymbolsPerSecond uint64
-	// reservation activation timestamp
-	StartTimestamp uint64
-	// reservation expiration timestamp
-	EndTimestamp uint64
-
-	// allowed quorums
-	QuorumNumbers []uint8
-	// ordered mapping of quorum number to payment split; on-chain validation should ensure split <= 100
-	QuorumSplits []byte
-}
-
-type OnDemandPayment struct {
-	// Total amount deposited by the user
-	CumulativePayment *big.Int
-}
-
-// PaymentQuorumConfig contains the configuration for a quorum's payment configurations
-// This is pretty much the same as the PaymentVaultTypesQuorumConfig struct in the contracts/bindings/IPaymentVault/binding.go file
-type PaymentQuorumConfig struct {
-	ReservationSymbolsPerSecond uint64
-
-	// OnDemand is initially only enabled on Quorum 0
-	OnDemandSymbolsPerSecond uint64
-	OnDemandPricePerSymbol   uint64
-}
-
-// PaymentQuorumProtocolConfig contains the configuration for a quorum's ratelimiting configurations
-// This is pretty much the same as the PaymentVaultTypesQuorumProtocolConfig struct in the contracts/bindings/IPaymentVault/binding.go file
-type PaymentQuorumProtocolConfig struct {
-	MinNumSymbols              uint64
-	ReservationAdvanceWindow   uint64
-	ReservationRateLimitWindow uint64
-
-	// OnDemand is initially only enabled on Quorum 0
-	OnDemandRateLimitWindow uint64
-	OnDemandEnabled         bool
-}
-
 type BlobVersionParameters struct {
 	// CodingRate specifies the amount of redundancy that will be added when encoding the blob
 	// (Note that for the purposes of integer representation, this is the inverse of the standard
@@ -688,19 +504,4 @@ type BlobVersionParameters struct {
 // and AdversaryThreshold that is valid for a given BlobVersionParameters.
 func (bvp *BlobVersionParameters) GetReconstructionThresholdBips() uint32 {
 	return RoundUpDivide(bvp.NumChunks*10000, (bvp.NumChunks-bvp.MaxNumOperators)*bvp.CodingRate)
-}
-
-// IsActive returns true if the reservation is active at the given timestamp
-func (ar *ReservedPayment) IsActive(currentTimestamp uint64) bool {
-	return WithinTime(time.Unix(int64(currentTimestamp), 0), time.Unix(int64(ar.StartTimestamp), 0), time.Unix(int64(ar.EndTimestamp), 0))
-}
-
-// IsActiveByNanosecond returns true if the reservation is active at the given timestamp
-func (ar *ReservedPayment) IsActiveByNanosecond(currentTimestamp int64) bool {
-	return WithinTime(time.Unix(0, currentTimestamp), time.Unix(int64(ar.StartTimestamp), 0), time.Unix(int64(ar.EndTimestamp), 0))
-}
-
-// WithinTime returns true if the timestamp is within the time range, inclusive of the start and end timestamps
-func WithinTime(timestamp time.Time, startTimestamp time.Time, endTimestamp time.Time) bool {
-	return !timestamp.Before(startTimestamp) && !timestamp.After(endTimestamp)
 }
