+++ core/meterer/meterer.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/Layr-Labs/eigenda/core"
-	"github.com/Layr-Labs/eigenda/core/meterer/paymentlogic"
+	"github.com/Layr-Labs/eigenda/core/payment"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 )
 
@@ -20,8 +20,6 @@ type Config struct {
 	UpdateInterval time.Duration
 }
 
-const OnDemandQuorumID = core.QuorumID(0)
-
 // Meterer handles payment accounting across different accounts. Disperser API server receives requests from clients and each request contains a blob header
 // with payments information (CumulativePayments, Timestamp, and Signature). Disperser will pass the blob header to the meterer, which will check if the
 // payments information is valid.
@@ -31,8 +29,8 @@ type Meterer struct {
 	// ChainPaymentState reads on-chain payment state periodically and caches it in memory
 	ChainPaymentState OnchainPayment
 
-	// MeteringStore tracks usage and payments in a storage backend
-	MeteringStore MeteringStore
+	// PaymentOffchainState tracks usage and payments in a storage backend
+	PaymentOffchainState payment.PaymentOffchainState
 
 	logger logging.Logger
 }
@@ -40,14 +38,14 @@ type Meterer struct {
 func NewMeterer(
 	config Config,
 	paymentChainState OnchainPayment,
-	meteringStore MeteringStore,
+	PaymentOffchainState payment.PaymentOffchainState,
 	logger logging.Logger,
 ) *Meterer {
 	return &Meterer{
-		Config:            config,
-		ChainPaymentState: paymentChainState,
-		MeteringStore:     meteringStore,
-		logger:            logger.With("component", "Meterer"),
+		Config:               config,
+		ChainPaymentState:    paymentChainState,
+		PaymentOffchainState: PaymentOffchainState,
+		logger:               logger.With("component", "Meterer"),
 	}
 }
 
@@ -73,7 +71,7 @@ func (m *Meterer) Start(ctx context.Context) {
 
 // MeterRequest validates a blob header and adds it to the meterer's state
 // TODO: return error if there's a rejection (with reasoning) or internal error (should be very rare)
-func (m *Meterer) MeterRequest(ctx context.Context, header core.PaymentMetadata, numSymbols uint64, quorumNumbers []uint8, receivedAt time.Time) (uint64, error) {
+func (m *Meterer) MeterRequest(ctx context.Context, header payment.PaymentMetadata, numSymbols uint64, quorumNumbers []uint8, receivedAt time.Time) (uint64, error) {
 	m.logger.Info("Validating incoming request's payment metadata", "paymentMetadata", header, "numSymbols", numSymbols, "quorumNumbers", quorumNumbers)
 
 	params, err := m.ChainPaymentState.GetPaymentGlobalParams()
@@ -81,7 +79,7 @@ func (m *Meterer) MeterRequest(ctx context.Context, header core.PaymentMetadata,
 		return 0, fmt.Errorf("failed to get payment global params: %w", err)
 	}
 	// Validate against the payment method
-	if !paymentlogic.IsOnDemandPayment(&header) {
+	if !payment.IsOnDemandPayment(&header) {
 		reservations, err := m.ChainPaymentState.GetReservedPaymentByAccountAndQuorums(ctx, header.AccountID, quorumNumbers)
 		if err != nil {
 			return 0, fmt.Errorf("failed to get active reservation by account: %w", err)
@@ -102,26 +100,26 @@ func (m *Meterer) MeterRequest(ctx context.Context, header core.PaymentMetadata,
 	// TODO(hopeyen): each quorum can have different min num symbols; the returned symbolsCharged is only for used for metrics.
 	// for now we simply return the charge for quorum 0, as quorums are likely to share the same min num symbols
 	// we can make this more granular by adding metrics to the meterer later on
-	_, protocolConfig, err := params.GetQuorumConfigs(OnDemandQuorumID)
+	_, protocolConfig, err := params.GetQuorumConfigs(payment.OnDemandDepositQuorumID)
 	if err != nil {
 		return 0, fmt.Errorf("failed to get on-demand quorum config: %w", err)
 	}
-	symbolsCharged := paymentlogic.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
+	symbolsCharged := payment.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
 	return symbolsCharged, nil
 }
 
 // serveReservationRequest handles the rate limiting logic for incoming requests
 func (m *Meterer) serveReservationRequest(
 	ctx context.Context,
-	globalParams *PaymentVaultParams,
-	header core.PaymentMetadata,
-	reservations map[core.QuorumID]*core.ReservedPayment,
+	globalParams *payment.PaymentVaultParams,
+	header payment.PaymentMetadata,
+	reservations map[core.QuorumID]*payment.ReservedPayment,
 	numSymbols uint64,
 	quorumNumbers []uint8,
 	receivedAt time.Time,
 ) error {
 	m.logger.Debug("Recording and validating reservation usage", "header", header, "reservation", reservations)
-	if err := paymentlogic.ValidateReservations(reservations, globalParams.QuorumProtocolConfigs, quorumNumbers, header.Timestamp, receivedAt.UnixNano()); err != nil {
+	if err := payment.ValidateReservations(reservations, globalParams.QuorumProtocolConfigs, quorumNumbers, header.Timestamp, receivedAt.UnixNano()); err != nil {
 		return fmt.Errorf("invalid reservation: %w", err)
 	}
 
@@ -134,9 +132,9 @@ func (m *Meterer) serveReservationRequest(
 
 // incrementBinUsage increments the bin usage atomically and checks for overflow
 func (m *Meterer) incrementBinUsage(
-	ctx context.Context, header core.PaymentMetadata,
-	reservations map[core.QuorumID]*core.ReservedPayment,
-	globalParams *PaymentVaultParams,
+	ctx context.Context, header payment.PaymentMetadata,
+	reservations map[core.QuorumID]*payment.ReservedPayment,
+	globalParams *payment.PaymentVaultParams,
 	numSymbols uint64,
 ) error {
 	charges := make(map[core.QuorumID]uint64)
@@ -148,15 +146,15 @@ func (m *Meterer) incrementBinUsage(
 		if err != nil {
 			return fmt.Errorf("failed to get quorum config for quorum %d: %w", quorumID, err)
 		}
-		charges[quorumID] = paymentlogic.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
+		charges[quorumID] = payment.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
 		quorumNumbers = append(quorumNumbers, quorumID)
 		reservationWindows[quorumID] = protocolConfig.ReservationRateLimitWindow
-		requestReservationPeriods[quorumID] = paymentlogic.GetReservationPeriodByNanosecond(header.Timestamp, protocolConfig.ReservationRateLimitWindow)
+		requestReservationPeriods[quorumID] = payment.GetReservationPeriodByNanosecond(header.Timestamp, protocolConfig.ReservationRateLimitWindow)
 	}
 	// Batch increment all quorums for the current quorums' reservation period
 	// For each quorum, increment by its specific symbolsCharged value
 	updatedUsages := make(map[core.QuorumID]uint64)
-	usage, err := m.MeteringStore.IncrementBinUsages(ctx, header.AccountID, quorumNumbers, requestReservationPeriods, charges)
+	usage, err := m.PaymentOffchainState.IncrementBinUsages(ctx, header.AccountID, quorumNumbers, requestReservationPeriods, charges)
 	if err != nil {
 		return err
 	}
@@ -169,7 +167,7 @@ func (m *Meterer) incrementBinUsage(
 	for quorumID, reservation := range reservations {
 		reservationWindow := reservationWindows[quorumID]
 		requestReservationPeriod := requestReservationPeriods[quorumID]
-		usageLimit := paymentlogic.GetBinLimit(reservation.SymbolsPerSecond, reservationWindow)
+		usageLimit := payment.GetBinLimit(reservation.SymbolsPerSecond, reservationWindow)
 		newUsage, ok := updatedUsages[quorumID]
 		if !ok {
 			return fmt.Errorf("failed to get updated usage for quorum %d", quorumID)
@@ -181,8 +179,8 @@ func (m *Meterer) incrementBinUsage(
 			// Bin was already filled before this increment
 			return fmt.Errorf("bin has already been filled for quorum %d", quorumID)
 		}
-		overflowPeriod := paymentlogic.GetOverflowPeriod(requestReservationPeriod, reservationWindow)
-		if charges[quorumID] <= usageLimit && overflowPeriod <= paymentlogic.GetReservationPeriod(int64(reservation.EndTimestamp), reservationWindow) {
+		overflowPeriod := payment.GetOverflowPeriod(requestReservationPeriod, reservationWindow)
+		if charges[quorumID] <= usageLimit && overflowPeriod <= payment.GetReservationPeriod(int64(reservation.EndTimestamp), reservationWindow) {
 			// Needs to go to overflow bin
 			overflowAmounts[quorumID] = newUsage - usageLimit
 			overflowPeriods[quorumID] = overflowPeriod
@@ -200,10 +198,10 @@ func (m *Meterer) incrementBinUsage(
 		for quorumID := range overflowAmounts {
 			overflowQuorums = append(overflowQuorums, quorumID)
 		}
-		_, err := m.MeteringStore.IncrementBinUsages(ctx, header.AccountID, overflowQuorums, overflowPeriods, overflowAmounts)
+		_, err := m.PaymentOffchainState.IncrementBinUsages(ctx, header.AccountID, overflowQuorums, overflowPeriods, overflowAmounts)
 		if err != nil {
 			// Rollback the increments for the current periods
-			rollbackErr := m.MeteringStore.DecrementBinUsages(ctx, header.AccountID, quorumNumbers, requestReservationPeriods, charges)
+			rollbackErr := m.PaymentOffchainState.DecrementBinUsages(ctx, header.AccountID, quorumNumbers, requestReservationPeriods, charges)
 			if rollbackErr != nil {
 				return fmt.Errorf("failed to increment overflow bins: %w; rollback also failed: %v", err, rollbackErr)
 			}
@@ -217,10 +215,10 @@ func (m *Meterer) incrementBinUsage(
 // serveOnDemandRequest handles the rate limiting logic for incoming requests
 // On-demand requests doesn't have additional quorum settings and should only be
 // allowed by ETH and EIGEN quorums
-func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *PaymentVaultParams, header core.PaymentMetadata, onDemandPayment *core.OnDemandPayment, symbolsCharged uint64, headerQuorums []uint8, receivedAt time.Time) error {
+func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *payment.PaymentVaultParams, header payment.PaymentMetadata, onDemandPayment *payment.OnDemandPayment, symbolsCharged uint64, headerQuorums []uint8, receivedAt time.Time) error {
 	m.logger.Debug("Recording and validating on-demand usage", "header", header, "onDemandPayment", onDemandPayment)
 
-	if err := paymentlogic.ValidateQuorum(headerQuorums, globalParams.OnDemandQuorumNumbers); err != nil {
+	if err := payment.ValidateQuorum(headerQuorums, globalParams.OnDemandQuorumNumbers); err != nil {
 		return fmt.Errorf("invalid quorum for On-Demand Request: %w", err)
 	}
 
@@ -229,14 +227,14 @@ func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *Paymen
 		return fmt.Errorf("request claims a cumulative payment greater than the on-chain deposit")
 	}
 
-	paymentConfig, protocolConfig, err := globalParams.GetQuorumConfigs(OnDemandQuorumID)
+	paymentConfig, protocolConfig, err := globalParams.GetQuorumConfigs(payment.OnDemandDepositQuorumID)
 	if err != nil {
 		return fmt.Errorf("failed to get payment config for on-demand quorum: %w", err)
 	}
 
-	symbolsCharged = paymentlogic.SymbolsCharged(symbolsCharged, protocolConfig.MinNumSymbols)
-	paymentCharged := paymentlogic.PaymentCharged(symbolsCharged, paymentConfig.OnDemandPricePerSymbol)
-	oldPayment, err := m.MeteringStore.AddOnDemandPayment(ctx, header, paymentCharged)
+	symbolsCharged = payment.SymbolsCharged(symbolsCharged, protocolConfig.MinNumSymbols)
+	paymentCharged := payment.PaymentCharged(symbolsCharged, paymentConfig.OnDemandPricePerSymbol)
+	oldPayment, err := m.PaymentOffchainState.AddOnDemandPayment(ctx, header, paymentCharged)
 	if err != nil {
 		return fmt.Errorf("failed to update cumulative payment: %w", err)
 	}
@@ -246,7 +244,7 @@ func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *Paymen
 		// If global bin usage update fails, roll back the payment to its previous value
 		// The rollback will only happen if the current payment value still matches what we just wrote
 		// This ensures we don't accidentally roll back a newer payment that might have been processed
-		dbErr := m.MeteringStore.RollbackOnDemandPayment(ctx, header.AccountID, header.CumulativePayment, oldPayment)
+		dbErr := m.PaymentOffchainState.RollbackOnDemandPayment(ctx, header.AccountID, header.CumulativePayment, oldPayment)
 		if dbErr != nil {
 			return dbErr
 		}
@@ -257,19 +255,19 @@ func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *Paymen
 }
 
 // IncrementGlobalBinUsage increments the bin usage atomically and checks for overflow
-func (m *Meterer) incrementGlobalBinUsage(ctx context.Context, params *PaymentVaultParams, symbolsCharged uint64, receivedAt time.Time) error {
-	paymentConfig, protocolConfig, err := params.GetQuorumConfigs(OnDemandQuorumID)
+func (m *Meterer) incrementGlobalBinUsage(ctx context.Context, params *payment.PaymentVaultParams, symbolsCharged uint64, receivedAt time.Time) error {
+	paymentConfig, protocolConfig, err := params.GetQuorumConfigs(payment.OnDemandDepositQuorumID)
 	if err != nil {
 		return fmt.Errorf("failed to get quorum configs for on-demand quorum: %w", err)
 	}
 
-	globalPeriod := paymentlogic.GetReservationPeriod(receivedAt.Unix(), protocolConfig.OnDemandRateLimitWindow)
+	globalPeriod := payment.GetReservationPeriod(receivedAt.Unix(), protocolConfig.OnDemandRateLimitWindow)
 
-	newUsage, err := m.MeteringStore.UpdateGlobalBin(ctx, globalPeriod, symbolsCharged)
+	newUsage, err := m.PaymentOffchainState.UpdateGlobalBin(ctx, globalPeriod, symbolsCharged)
 	if err != nil {
 		return fmt.Errorf("failed to increment global bin usage: %w", err)
 	}
-	if newUsage > paymentlogic.GetBinLimit(paymentConfig.OnDemandSymbolsPerSecond, protocolConfig.OnDemandRateLimitWindow) {
+	if newUsage > payment.GetBinLimit(paymentConfig.OnDemandSymbolsPerSecond, protocolConfig.OnDemandRateLimitWindow) {
 		return fmt.Errorf("global bin usage overflows")
 	}
 	return nil
