+++ core/mock/payment_state.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Layr-Labs/eigenda/core"
 	"github.com/Layr-Labs/eigenda/core/meterer"
+	"github.com/Layr-Labs/eigenda/core/payment"
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/mock"
 )
@@ -20,13 +21,13 @@ func (m *MockOnchainPaymentState) RefreshOnchainPaymentState(ctx context.Context
 	return args.Error(0)
 }
 
-func (m *MockOnchainPaymentState) GetReservedPaymentByAccountAndQuorums(ctx context.Context, accountID gethcommon.Address, quorumNumbers []core.QuorumID) (map[core.QuorumID]*core.ReservedPayment, error) {
+func (m *MockOnchainPaymentState) GetReservedPaymentByAccountAndQuorums(ctx context.Context, accountID gethcommon.Address, quorumNumbers []core.QuorumID) (map[core.QuorumID]*payment.ReservedPayment, error) {
 	args := m.Called(ctx, accountID, quorumNumbers)
-	var value map[core.QuorumID]*core.ReservedPayment
-	if fn, ok := args.Get(0).(func(context.Context, gethcommon.Address, []core.QuorumID) map[core.QuorumID]*core.ReservedPayment); ok {
+	var value map[core.QuorumID]*payment.ReservedPayment
+	if fn, ok := args.Get(0).(func(context.Context, gethcommon.Address, []core.QuorumID) map[core.QuorumID]*payment.ReservedPayment); ok {
 		value = fn(ctx, accountID, quorumNumbers)
 	} else if args.Get(0) != nil {
-		value = args.Get(0).(map[core.QuorumID]*core.ReservedPayment)
+		value = args.Get(0).(map[core.QuorumID]*payment.ReservedPayment)
 	}
 	var err error
 	if len(args) > 1 {
@@ -35,11 +36,11 @@ func (m *MockOnchainPaymentState) GetReservedPaymentByAccountAndQuorums(ctx cont
 	return value, err
 }
 
-func (m *MockOnchainPaymentState) GetOnDemandPaymentByAccount(ctx context.Context, accountID gethcommon.Address) (*core.OnDemandPayment, error) {
+func (m *MockOnchainPaymentState) GetOnDemandPaymentByAccount(ctx context.Context, accountID gethcommon.Address) (*payment.OnDemandPayment, error) {
 	args := m.Called(ctx, accountID)
-	var value *core.OnDemandPayment
+	var value *payment.OnDemandPayment
 	if args.Get(0) != nil {
-		value = args.Get(0).(*core.OnDemandPayment)
+		value = args.Get(0).(*payment.OnDemandPayment)
 	}
 	return value, args.Error(1)
 }
@@ -53,11 +54,11 @@ func (m *MockOnchainPaymentState) GetQuorumNumbers(ctx context.Context) ([]core.
 	return value, args.Error(1)
 }
 
-func (m *MockOnchainPaymentState) GetPaymentGlobalParams() (*meterer.PaymentVaultParams, error) {
+func (m *MockOnchainPaymentState) GetPaymentGlobalParams() (*payment.PaymentVaultParams, error) {
 	args := m.Called()
-	var value *meterer.PaymentVaultParams
+	var value *payment.PaymentVaultParams
 	if args.Get(0) != nil {
-		value = args.Get(0).(*meterer.PaymentVaultParams)
+		value = args.Get(0).(*payment.PaymentVaultParams)
 	}
 	return value, args.Error(1)
 }
